% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ares-package.R
\docType{package}
\name{ares-package}
\alias{ares-package}
\title{ares: Fast Multivariate Adaptive Regression Splines}
\description{
\code{ares} fits MARS models using Friedman's (1991) fast least-squares forward
selection and a parallelized knot search built on \code{Rcpp} and \code{RcppParallel}.
Designed for numerical parity with the \code{earth} package on the gaussian-only
core while taking advantage of multi-core CPUs.
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines.
\emph{Annals of Statistics}, 19(1):1-67.
}
\author{
\strong{Maintainer}: Jack Trametta \email{jtrametta@gmail.com}

Authors:
\itemize{
  \item Jack Trametta \email{jtrametta@gmail.com}
}

}
\keyword{internal}
