% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.plda}
\alias{plot.plda}
\title{Projection plot for \code{plda} fits}
\usage{
\method{plot}{plda}(x, data = NULL, labels = NULL, ...)
}
\arguments{
\item{x}{A \code{"plda"} object.}

\item{data}{Predictor matrix (required).}

\item{labels}{Optional class label vector for colouring points.}

\item{...}{Further graphical parameters passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Projection plot for \code{plda} fits
}
