% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.plda}
\alias{print.plda}
\title{Print method for \code{plda} fits}
\usage{
\method{print}{plda}(x, ...)
}
\arguments{
\item{x}{A \code{"plda"} object.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{plda} fits
}
\examples{
fit <- plda(Species ~ ., data = iris, lambda = 0.1, autotune = FALSE)
print(fit)
}
