% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.ares}
\alias{summary.ares}
\alias{print.summary.ares}
\title{Summary method for \code{ares} fits}
\usage{
\method{summary}{ares}(object, ...)

\method{print}{summary.ares}(x, ...)
}
\arguments{
\item{object}{an \code{ares} object}

\item{...}{unused}

\item{x}{a \code{summary.ares} object}
}
\value{
An object of class \code{"summary.ares"}.
}
\description{
Summary method for \code{ares} fits
}
\examples{
fit <- ares(as.matrix(mtcars[, -1]), mtcars$mpg, nthreads = 2)
summary(fit)
}
