% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ares}
\alias{predict.ares}
\title{Predictions from an \code{ares} fit}
\usage{
\method{predict}{ares}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"ares"}.}

\item{newdata}{A numeric matrix or data frame of new predictors. If \code{NULL}
(default), returns \code{object$fitted.values}.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
A numeric vector of length \code{nrow(newdata)}.
}
\description{
Predictions from an \code{ares} fit
}
\examples{
fit <- ares(as.matrix(mtcars[, -1]), mtcars$mpg, nthreads = 2)
p <- predict(fit, as.matrix(mtcars[, -1]))
}
