% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ares-package.R
\docType{package}
\name{roadrunner-package}
\alias{roadrunner}
\alias{roadrunner-package}
\title{roadrunner: Fast, Low-Dependency Machine Learning Algorithms}
\description{
\code{roadrunner} is a collection of fast, low-dependency implementations of
classical machine learning algorithms. The first algorithm exposed is
\code{\link[=ares]{ares()}}, a Multivariate Adaptive Regression Splines (MARS) fitter built
on \code{Rcpp} and \code{RcppParallel}. Additional algorithms will be added as
separate exported functions under the same package roof.
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines.
\emph{Annals of Statistics}, 19(1):1-67.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cetialphafive.github.io/roadrunner/}
  \item \url{https://github.com/CetiAlphaFive/roadrunner}
  \item Report bugs at \url{https://github.com/CetiAlphaFive/roadrunner/issues}
}

}
\author{
\strong{Maintainer}: Jack T. Rametta \email{jtrametta@gmail.com}

Authors:
\itemize{
  \item Jack T. Rametta \email{jtrametta@gmail.com}
}

}
\keyword{internal}
