% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.plda}
\alias{summary.plda}
\alias{print.summary.plda}
\title{Summary method for \code{plda} fits}
\usage{
\method{summary}{plda}(object, ...)

\method{print}{summary.plda}(x, ...)
}
\arguments{
\item{object}{A \code{"plda"} object.}

\item{...}{Currently ignored.}

\item{x}{A \code{summary.plda} object.}
}
\value{
An object of class \code{"summary.plda"}.

Invisibly returns \code{x}.
}
\description{
Summary method for \code{plda} fits
}
\examples{
fit <- plda(Species ~ ., data = iris, lambda = 0.1, autotune = FALSE)
summary(fit)
fit <- plda(Species ~ ., data = iris, lambda = 0.1, autotune = FALSE)
print(summary(fit))
}
